requirements.txt
dj-database-url==0.4.1
Django==1.10.4
django-crispy-forms==1.6.1
django-filter==1.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.2
